app_2/requirements.txt
gunicorn==20.1.0
fastapi==0.103.2
uvicorn==0.24.0
asyncio==3.4.3
httpx==0.26.0
aiohttp==3.9.1
pydantic==2.5.3
python-dotenv==1.0.1
loguru==0.7.2
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0